TIP-860: Move Currency related api tests
